requirements.txt
flask
matplotlib
numpy
pillow
scipy